refactor(internal/librarian/python): move package up one level
Align the Python package structure with internal/librarian/rust by
moving it up one directory level.
